third_party/WebKit/LayoutTests/external/wpt/referrer-policy/strict-origin/attr-referrer/cross-origin/http-http/script-tag/insecure-protocol.keep-origin-redirect.http-expected.txt

This is a testharness.js-based test.
FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is cross-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://web-platform.test:8001/" but got "http://web-platform.test:8001/referrer-policy/strict-origin/attr-referrer/cross-origin/http-http/script-tag/insecure-protocol.keep-origin-redirect.http.html"
Harness: the test ran to completion.
